Better caching support (#37)
- Moved caching to a backend controlled option.
- Added the ability to test-run changes via the `dry-run` keyword.
- Fixed `get_service_enabled` function such that both v1 and v2 formats are correctly handled.
- Added v1 support to the `get_service` function.
- Upgrade click_log and moved to a application configuration object for CLI.
- Added a `list_service` command to CLI.
- Added a `deploy_service` command to CLI.
- Added changelog.

+++ swag_client/backends/dynamodb.py
@@ -1,11 +1,13 @@
 import logging
 
 import boto3
-import jmespath
+from dogpile.cache import make_region
 
 from swag_client.backend import SWAGManager
 
 logger = logging.getLogger(__file__)
+
+dynamodb_region = make_region()
 
 
 class DynamoDBSWAGManager(SWAGManager):
@@ -15,61 +17,53 @@ class DynamoDBSWAGManager(SWAGManager):
         resource = boto3.resource('dynamodb', region_name=kwargs['region'])
         self.table = resource.Table(namespace)
 
-    def get(self, search_filter):
-        """Fetch one item from file. Applying given filter, raises exception if more than one item found."""
-        logger.debug('Fetching items. Filter: {filter}. Table: {namespace}'.format(
-            filter=search_filter,
-            namespace=self.namespace
-        ))
+        if not dynamodb_region.is_configured:
+            dynamodb_region.configure(
+                'dogpile.cache.memory',
+                expiration_time=kwargs['cache_expires']
+            )
 
-        items = self.table.scan()['Items']
-        results = jmespath.search(search_filter, items)
-
-        if len(results) > 1:
-            raise Exception('Found more than one result for get!')
-
-        elif len(results) == 1:
-            return results[0]
-
-    def create(self, item):
+    def create(self, item, dry_run=None):
         """Creates a new item in file."""
         logger.debug('Creating new item. Item: {item} Table: {namespace}'.format(
             item=item,
             namespace=self.namespace
         ))
 
-        self.table.put_item(Item=item)
+        if not dry_run:
+            self.table.put_item(Item=item)
 
         return item
 
-    def delete(self, item):
+    def delete(self, item, dry_run=None):
         """Deletes item in file."""
         logger.debug('Deleting item. Item: {item} Table: {namespace}'.format(
             item=item,
             namespace=self.namespace
         ))
 
-        self.table.delete_item(Key={'id': item['id']})
+        if not dry_run:
+            self.table.delete_item(Key={'id': item['id']})
 
-    def update(self, item):
+        return item
+
+    def update(self, item, dry_run=None):
         """Updates item info in file."""
         logger.debug('Updating item. Item: {item} Table: {namespace}'.format(
             item=item,
             namespace=self.namespace
         ))
-        self.table.put_item(Item=item)
+
+        if not dry_run:
+            self.table.put_item(Item=item)
+
         return item
 
-    def get_all(self, search_filter=None):
+    @dynamodb_region.cache_on_arguments()
+    def get_all(self):
         """Gets all items in file."""
-        logger.debug('Fetching items. Filter: {filter}. Table: {namespace}'.format(
-            filter=search_filter,
+        logger.debug('Fetching items. Table: {namespace}'.format(
             namespace=self.namespace
         ))
 
-        items = self.table.scan()['Items']
-
-        if search_filter:
-            return jmespath.search(search_filter, items)
-
-        return items
+        return self.table.scan()['Items']
