+++ swag_client/backends/file.py
@@ -4,9 +4,10 @@ import json
 import logging
 from io import open
 
-import jmespath
+from dogpile.cache import make_region
 
 from swag_client.backend import SWAGManager
+from swag_client.util import append_item, remove_item
 
 logger = logging.getLogger(__name__)
 
@@ -14,6 +15,9 @@ try:
     from json.decoder import JSONDecodeError
 except ImportError:
     JSONDecodeError = ValueError
+
+
+file_region = make_region()
 
 
 def load_file(data_file):
@@ -26,8 +30,11 @@ def load_file(data_file):
         return []
 
 
-def save_file(data_file, data):
+def save_file(data_file, data, dry_run=None):
     """Writes JSON data to data file."""
+    if dry_run:
+        return
+
     with open(data_file, 'w', encoding='utf-8') as f:
         if sys.version_info > (3, 0):
             f.write(json.dumps(data))
@@ -41,6 +48,12 @@ class FileSWAGManager(SWAGManager):
         self.namespace = namespace
         self.version = kwargs['schema_version']
 
+        if not file_region.is_configured:
+            file_region.configure(
+                'dogpile.cache.memory',
+                expiration_time=kwargs['cache_expires']
+            )
+
         if not kwargs.get('data_file'):
             self.data_file = os.path.join(kwargs['data_dir'], self.namespace + '.json')
         else:
@@ -53,86 +66,46 @@ class FileSWAGManager(SWAGManager):
 
             save_file(self.data_file, [])
 
-    def get(self, search_filter):
-        """Fetch one item from file. Applying given filter, raises exception if more than one item found."""
-        logger.debug('Fetching items. Filter: {filter}. Path: {data_file}'.format(
-            filter=search_filter,
-            data_file=self.data_file
-        ))
-
-        items = load_file(self.data_file)
-        results = jmespath.search(search_filter, items)
-
-        if len(results) > 1:
-            raise Exception('Found more than one result for get!')
-
-        elif len(results) == 1:
-            return results[0]
-
-    def create(self, item):
+    def create(self, item, dry_run=None):
         """Creates a new item in file."""
         logger.debug('Creating new item. Item: {item} Path: {data_file}'.format(
             item=item,
             data_file=self.data_file
         ))
 
-        items = self.get_all()
-        if self.version == 1:
-            items[self.namespace].append(item)
-        else:
-            items.append(item)
+        items = load_file(self.data_file)
+        items = append_item(self.namespace, self.version, item, items)
+        save_file(self.data_file, items, dry_run=dry_run)
 
-        save_file(self.data_file, items)
         return item
 
-    def delete(self, item):
+    def delete(self, item, dry_run=None):
         """Deletes item in file."""
         logger.debug('Deleting item. Item: {item} Path: {data_file}'.format(
             item=item,
             data_file=self.data_file
         ))
 
-        if self.version == 1:
-            # TODO only supports aws providers
-            items = self.get_all("{namespace}[?id!='aws-{id}']".format(
-                id=item['id'],
-                namespace=self.namespace
-            ))
-        else:
-            items = self.get_all("[?id!='{id}']".format(id=item['id']))
+        items = load_file(self.data_file)
+        items = remove_item(self.namespace, self.version, item, items)
+        save_file(self.data_file, items, dry_run=dry_run)
 
-        save_file(self.data_file, items)
+        return item
 
-    def update(self, item):
+    def update(self, item, dry_run=None):
         """Updates item info in file."""
         logger.debug('Updating item. Item: {item} Path: {data_file}'.format(
             item=item,
             data_file=self.data_file
         ))
-        self.delete(item)
-        return self.create(item)
+        self.delete(item, dry_run=dry_run)
+        return self.create(item, dry_run=dry_run)
 
-    def get_all(self, search_filter=None):
+    @file_region.cache_on_arguments()
+    def get_all(self):
         """Gets all items in file."""
-        logger.debug('Fetching items. Filter: {filter}. Path: {data_file}'.format(
-            filter=search_filter,
+        logger.debug('Fetching items. Path: {data_file}'.format(
             data_file=self.data_file
         ))
 
-        items = load_file(self.data_file)
-
-        # default values
-        if not items:
-            if self.version == 1:
-                return {self.namespace: []}
-            return []
-
-        # filter values
-        if search_filter:
-            items = jmespath.search(search_filter, items)
-            if self.version == 1:
-                return {self.namespace: items}
-            return items
-
-        # return raw values
-        return items
+        return load_file(self.data_file)
